src/website/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.38.0
dj-database-url==2.1.0
Django==5.0.7
django-author==1.2.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-modeltranslation==0.18.11
django-nested-admin==4.0.2
django-quill-editor==0.1.40
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
invoke==2.2.0
packaging==23.2
polib==1.2.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.1
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools
setuptools-git==1.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
